+++ src/lib/Offcanvas/tests/offcanvas-navbar-test.svelte
@@ -1,22 +1,19 @@
 <script lang="ts">
     import { Container, Navbar, Offcanvas } from '$lib/index.js';
 
-    let { onShown, onHidden }: Pick<Offcanvas.RootProps, 'onShown' | 'onHidden'> = $props();
-    let isShown = $state(false);
+    let {
+        onShown,
+        onHidden,
+        useBackdrop = false,
+        showOnBreakpoint,
+        isShown
+    }: Pick<Offcanvas.RootProps, 'onShown' | 'onHidden' | 'useBackdrop' | 'showOnBreakpoint' | 'isShown'> = $props();
 </script>
 
 <Navbar.Root expandOnBreakpoint="lg">
     <Container>
         <Navbar.Toggler data-testid="navbar-toggler"><Navbar.TogglerIcon /></Navbar.Toggler>
-        <Offcanvas.Root
-            {isShown}
-            useBackdrop={false}
-            {onShown}
-            onHidden={(event) => {
-                isShown = false;
-                onHidden?.(event);
-            }}
-            data-testid="navbar-offcanvas">
+        <Offcanvas.Root {isShown} {useBackdrop} {showOnBreakpoint} {onShown} {onHidden} data-testid="navbar-offcanvas">
             <Offcanvas.Header isDismissible={true}><Offcanvas.Title>Menu</Offcanvas.Title></Offcanvas.Header>
             <Offcanvas.Body>Body</Offcanvas.Body>
         </Offcanvas.Root>